refactor: 최종 금액 구하는 로직을 MatchResult가 담당한다
