+++ src/test/java/blackjack/model/result/MatchResultTest.java
@@ -1,8 +1,8 @@
-package blackjack.model.referee;
+package blackjack.model.result;
 
 import blackjack.model.card.Score;
-import blackjack.model.result.MatchResult;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -32,5 +32,31 @@ class MatchResultTest {
                 Arguments.of(Score.from(8), MatchResult.LOSE),
                 Arguments.of(Score.from(9), MatchResult.PUSH)
         );
+    }
+
+    @Test
+    @DisplayName("플레이어가 승리하면 베팅 금액만큼 금액을 더 받는다")
+    void determineWinnerFinalBettingMoneyTest() {
+        // given
+        BettingMoney bettingMoney = BettingMoney.from(1000);
+
+        // when
+        BettingMoney finalBettingMoney = MatchResult.WIN.calculateFinalMoney(bettingMoney);
+
+        // then
+        assertThat(finalBettingMoney.getAmount()).isEqualTo(2000);
+    }
+
+    @Test
+    @DisplayName("플레이어가 지면 베팅 금액을 잃는다")
+    void determineLoserFinalBettingMoneyTest() {
+        // given
+        BettingMoney bettingMoney = BettingMoney.from(1000);
+
+        // when
+        BettingMoney finalBettingMoney = MatchResult.LOSE.calculateFinalMoney(bettingMoney);
+
+        // then
+        assertThat(finalBettingMoney.getAmount()).isEqualTo(0);
     }
 }
